Connect to Spilo Masters or InstanceIDs
When not specifying a dns-name or IP address, we assume
we are looking for a named Spilo.
If so, we use boto3 to determine the IP address of the remote host
to use for tunneling.

As we are using boto3, this also allows us to query multiple regions.
If a named Spilo connection is requested, it will search all specified
(or default) regions for a Spilo master.