Ignore SIGINT while in command mode
When Gotthard is running a command, we want only the command
that is actually running to handle the interrupt.

We also do not want the interrupto to go to the ssh tunnel itself,
it should not be terminated if a SIGINT is sent.

For the command running we use a thread to accomplish this goal.

Resolves issue #2